requirements.txt
yt-dlp==2024.4.9
discord==2.3.2
python-dotenv==1.0.1
PyNaCl==1.5.0